Fix for crash when storing attribute as SEGYDirect
Summary:
- Plus some fixes related to storing attribute as ZGY.
- Plus changes in text when importing SEGY as SEGYDirect

Jira: OD-4236

Test Plan: No crash anymore

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15334

+++ src/Seis/segydirecttr.cc
@@ -673,9 +673,12 @@ bool SEGYDirectSeisTrcTranslator::implRemove( const IOObj* ioobj,
     if ( deep )
     {
 	SEGY::DirectDef segydef( ioobj->mainFileName() );
-	const SEGY::FileDataSet& fds = segydef.fileDataSet();
-	for ( int idx=0; idx<fds.nrFiles(); idx++ )
-	    File::remove( fds.fileName(idx) );
+	if ( !segydef.isEmpty() )
+	{
+	    const SEGY::FileDataSet& fds = segydef.fileDataSet();
+	    for ( int idx=0; idx<fds.nrFiles(); idx++ )
+		File::remove( fds.fileName(idx) );
+	}
     }
 
     Translator::implRemove( ioobj );
